+++ sqlite3_web/test/client_test.dart
@@ -16,7 +16,6 @@ void main() {
       DatabaseImplementation.opfsShared,
       DatabaseImplementation.opfsWithExternalLocks,
       DatabaseImplementation.opfsWithExternalLocksWorkaround,
-      DatabaseImplementation.opfsAtomics,
       DatabaseImplementation.indexedDbShared,
       DatabaseImplementation.indexedDbUnsafeWorker,
       DatabaseImplementation.indexedDbUnsafeLocal,